# backend/openapi.yaml
openapi: 3.1.0
info:
  title: FlashFound Backend API
  version: 0.6.0
  description: FlashFound backend API documented through Phase 6.
servers:
  - url: http://127.0.0.1:8080
    description: Local development
security:
  - bearerAuth: []
paths:
...
  /api/v1/profile/me:
    get:
      summary: Get my profile
      responses:
        "200":
          description: Profile
    patch:
      summary: Patch profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                display_name:
                  type: string
                display_avatar_url:
                  type: string
                  format: uri
                  nullable: true
      responses:
        "200":
          description: Updated profile
        "400":
          $ref: "#/components/responses/Error"
  /api/v1/profile/verification-selfie:
    patch:
      summary: Update verification selfie
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [verification_selfie_url]
              properties:
                verification_selfie_url:
                  type: string
                  format: uri
                face_verification_completed:
                  type: boolean
      responses:
        "200":
          description: Updated profile
...
  /api/v1/verification/session/{id}:
    get:
      summary: Get verification session
      parameters:
        - $ref: "#/components/parameters/RequestId"
      responses:
        "200":
          description: Session state
  /api/v1/verification/session/{id}/submit:
    post:
      summary: Submit selfie URL for verification
      parameters:
        - $ref: "#/components/parameters/RequestId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [selfie_url]
              properties:
                selfie_url:
                  type: string
                  format: uri
      responses:
        "200":
          description: Submission result
  /api/v1/verification/session/{id}/finalize:
    post:
      summary: Finalize verification session
      parameters:
        - $ref: "#/components/parameters/RequestId"
      responses:
        "200":
          description: Profile verified
...
  /api/v1/private-access/requests:
    post:
      summary: Create private access request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [event_id, target_user_id]
              properties:
                event_id:
                  type: string
                  format: uuid
                target_user_id:
                  type: string
                  format: uuid
      responses:
        "200":
          description: Request created
        "429":
          $ref: "#/components/responses/Error"
  /api/v1/private-access/requests/{id}/approve:
    post:
      summary: Approve private access request
      parameters:
        - $ref: "#/components/parameters/RequestId"
      responses:
        "200":
          description: Approved
  /api/v1/private-access/requests/{id}/deny:
    post:
      summary: Deny private access request
      parameters:
        - $ref: "#/components/parameters/RequestId"
      responses:
        "200":
          description: Denied
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    adminBearer:
      type: http
      scheme: bearer
      bearerFormat: FlashFoundAdminToken
  parameters:
...
    RequestId:
      in: path
      name: id
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Error:
      description: Structured error envelope
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
...
    ErrorEnvelope:
      type: object
      properties:
        ok:
          type: boolean
          const: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            requestId:
              type: string
            details:
              nullable: true
...